cephadm: create host_facts.py in cephadmlib
For storing classes/functions related to gathering
information about the hosts such as disk enclosures
and networks
